pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "gradio>=5.25.2",
    "langchain-google-genai>=2.1.3",
    "mcp-use>=1.2.5",
    "mcp[cli]>=1.6.0",
    "praisonaiagents[llm]>=0.0.76",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
